+++ src/core/model/Hero.java
@@ -2,7 +2,7 @@ package core.model;
 
 public class Hero extends GameObject {
 
-    public Hero(float x, float y, float direction, float speed) {
+    public Hero(double x, double y, double direction, double speed) {
         super(x, y, direction, speed);
     }
